services/embedding_worker/requirements.txt
nats-py
httpx
sqlalchemy[asyncio]
asyncpg
sqlalchemy-vector
python-dotenv

